requirements.txt
Django>=4.2
djangorestframework
djangorestframework-simplejwt
requests
python-dotenv
transformers
torch
